removed several bad casts to 'long'.
The 'long' type should generally be avoided because its size differs on Windows and Linux/Mac so it is almost guaranteed to cause undefined behavior

+++ src/common/rendering/hwrenderer/data/hw_shaderpatcher.cpp
@@ -246,7 +246,7 @@ FString RemoveLayoutLocationDecl(FString code, const char *inoutkeyword)
 
 		// keyword following the declaration?
 		bool keywordFound = true;
-		long i;
+		ptrdiff_t i;
 		for (i = 0; inoutkeyword[i] != 0; i++)
 		{
 			if (chars[endIndex + i] != inoutkeyword[i])
